Using native splash screen instead of custom splash screen
