+++ android/app/build.gradle
@@ -202,6 +202,7 @@ dependencies {
     }
 
     implementation("com.google.android.gms:play-services-auth:20.4.1")
+    implementation("androidx.core:core-splashscreen:1.0.0")
 }
 
 apply from: file("../../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesAppBuildGradle(project)
