+++ android/app/src/main/res/values/colors.xml
@@ -1,0 +1,3 @@
+<resources>
+    <color name="bootsplash_background">#FFFFFF</color>
+</resources>
