+++ package.json
@@ -34,6 +34,7 @@
     "react-hook-form": "^7.43.8",
     "react-native": "0.71.4",
     "react-native-animatable": "^1.3.3",
+    "react-native-bootsplash": "^4.7.3",
     "react-native-calendars": "^1.1298.0",
     "react-native-collapsible": "^1.6.1",
     "react-native-config": "^1.5.0",
