+++ src/app/navigation/RootNavigator.tsx
@@ -3,9 +3,10 @@ import {
   createNativeStackNavigator,
   NativeStackNavigationOptions,
 } from '@react-navigation/native-stack';
+import RNBootSplash from 'react-native-bootsplash';
+import {DefaultTheme, NavigationContainer} from '@react-navigation/native';
 
 import LoginScreen from '../screens/LoginScreen';
-import SplashScreen from '../screens/SplashScreen';
 import TimesheetList from '../screens/TimesheetScreen/view/timesheetList';
 import ProfileScreen from '../screens/ProfileScreen';
 import LeaveDetailScreen from '../screens/LeaveScreen/ManagementLeaveScreen/LeaveDetailScreen';
@@ -14,6 +15,7 @@ import DrawerNavigator from './DrawerNavigation';
 import UserContext from '../context/user.context';
 import AsyncStore from '../services/asyncStorage';
 import {initNotificationService} from '../services/firebase/messaging';
+import {navigationRef} from '.';
 
 import {RootStackParamList} from './types';
 import {
@@ -23,11 +25,20 @@ import {
   USER_PROFILE_SCREEN,
   USER_TIMESHEET,
 } from '../constant/screenNames';
+import colors from '../constant/colors';
 
 const RootStack = createNativeStackNavigator<RootStackParamList>();
 
 const screenOptions: NativeStackNavigationOptions = {
   headerShown: false,
+};
+
+const theme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    background: colors.WHITE,
+  },
 };
 
 const RootNavigator = () => {
@@ -47,7 +58,6 @@ const RootNavigator = () => {
         setUserContextData({authToken, userData: JSON.parse(userData)});
       }
 
-      await new Promise<void>(resolve => setTimeout(resolve, 1000));
       setLoading(false);
     };
 
@@ -55,31 +65,36 @@ const RootNavigator = () => {
   }, [setUserContextData]);
 
   if (loading) {
-    return <SplashScreen />;
+    return null;
   }
 
   return (
-    <RootStack.Navigator
-      screenOptions={screenOptions}
-      initialRouteName={DRAWER}>
-      {userContextData ? (
-        <>
-          <RootStack.Screen name={DRAWER} component={DrawerNavigator} />
+    <NavigationContainer
+      theme={theme}
+      ref={navigationRef}
+      onReady={() => RNBootSplash.hide({fade: true})}>
+      <RootStack.Navigator
+        screenOptions={screenOptions}
+        initialRouteName={DRAWER}>
+        {userContextData ? (
+          <>
+            <RootStack.Screen name={DRAWER} component={DrawerNavigator} />
 
-          <RootStack.Screen name={USER_TIMESHEET} component={TimesheetList} />
-          <RootStack.Screen
-            name={USER_PROFILE_SCREEN}
-            component={ProfileScreen}
-          />
-          <RootStack.Screen
-            name={LEAVE_DETAIL_SCREEN}
-            component={LeaveDetailScreen}
-          />
-        </>
-      ) : (
-        <RootStack.Screen name={LOGIN_SCREEN} component={LoginScreen} />
-      )}
-    </RootStack.Navigator>
+            <RootStack.Screen name={USER_TIMESHEET} component={TimesheetList} />
+            <RootStack.Screen
+              name={USER_PROFILE_SCREEN}
+              component={ProfileScreen}
+            />
+            <RootStack.Screen
+              name={LEAVE_DETAIL_SCREEN}
+              component={LeaveDetailScreen}
+            />
+          </>
+        ) : (
+          <RootStack.Screen name={LOGIN_SCREEN} component={LoginScreen} />
+        )}
+      </RootStack.Navigator>
+    </NavigationContainer>
   );
 };
